update alert query moved to extension method
