+++ src/Application/Common/Extensions/SiteWatcherContextExtensions.cs
@@ -1,6 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using SiteWatcher.Application.Interfaces;
+using SiteWatcher.Domain.Alerts;
+using SiteWatcher.Domain.Alerts.DTOs;
 using SiteWatcher.Domain.Alerts.Enums;
+using SiteWatcher.Domain.Common.ValueObjects;
 using SiteWatcher.Domain.Users;
 
 namespace SiteWatcher.Application.Common.Extensions;
@@ -28,5 +31,33 @@ public static class SiteWatcherContextExtensions
             .ThenInclude(a => a.Rule)
             .Take(take)
             .ToArrayAsync(ct);
+    }
+
+    public static async Task<Alert?> GetAlertForUpdateAsync(this ISiteWatcherContext ctx, AlertId alertId, UserId userId,
+        CancellationToken ct)
+    {
+        var alertDto = await ctx.Alerts
+            .Where(a => a.Id == alertId && a.UserId == userId && a.Active)
+            .Select(alert => new UpdateAlertDto
+            {
+                Id = alert.Id,
+                UserId = alert.UserId,
+                CreatedAt = alert.CreatedAt,
+                Name = alert.Name,
+                Frequency = alert.Frequency,
+                SiteName = alert.Site.Name,
+                SiteUri = alert.Site.Uri,
+                Rule = alert.Rule,
+                LastVerification = alert.LastVerification
+            })
+            .SingleOrDefaultAsync(ct);
+
+        if (alertDto is null)
+            return null;
+
+        var alert = Alert.GetModelForUpdate(alertDto);
+        // Attach alert and site, because the rule is already tracked
+        ctx.Attach(alert);
+        return alert;
     }
 }
